# full-rag-springboot-sync/target/classes/static/openapi.yaml
openapi: 3.0.0
info:
  title: Full RAG Spring Boot Sync API
  description: |
    Retrieval Augmented Generation API for document ingestion and querying.
    This API provides endpoints to ingest documents into a vector database and
    query them using natural language questions with AI-powered answers.
  version: 1.0.0
  contact:
    name: RAG API Support
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/rag/ingest:
    post:
      summary: Ingest a document into the RAG system
      description: |
        Ingests a text document by chunking, embedding, and storing in vector database.
        Supports long documents with automatic chunking and overlap.
        The document will be split into chunks, each chunk will be embedded using
        the configured embedding model, and stored in Qdrant vector database.
      operationId: ingestDocument
      tags:
        - RAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - docId
                - text
              properties:
                docId:
                  type: string
                  description: Unique identifier for the document
                  example: "user-manual-v1"
                  minLength: 1
                text:
                  type: string
                  description: Full text content of the document
                  example: "This is a comprehensive user manual..."
                  minLength: 10
      responses:
        '200':
          description: Document ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  docId:
                    type: string
                    example: "user-manual-v1"
                  chunksCreated:
                    type: integer
                    example: 15
                    description: Number of chunks created from the document
                  message:
                    type: string
                    example: "Document 'user-manual-v1' ingested successfully with 15 chunks"
        '400':
          description: Invalid request (missing fields, empty text)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  docId:
                    type: string
                  chunksCreated:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: "Validation error: text cannot be null or empty"
        '500':
          description: Server error during ingestion
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  docId:
                    type: string
                  chunksCreated:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: "Internal server error: ..."
...